Dev Profile

+++ client/include/GUI/Audio/ISoundManager.hh
@@ -14,8 +14,8 @@ namespace Gui
 
 
       virtual void	loadSound(const std::string &, const std::string &) = 0;
-      virtual void	playSound(const std::string &, int = 50) = 0;
-      virtual void	playMusic(const std::string &, int = 50) = 0;
+      virtual void	playSound(const std::string &, float = 50) = 0;
+      virtual void	playMusic(const std::string &, float = 50) = 0;
       virtual void	stopMusic() = 0;
     };
   }
